examples/python/requirements.txt
pyyaml>=6.0
flask>=2.3.0
pytest>=7.4.0
